% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC.CO_Evaluation.R
\name{DC.CO_Evaluation}
\alias{DC.CO_Evaluation}
\title{DC.CO_Evaluation}
\usage{
DC.CO_Evaluation(
  mutation_data = mutation_data,
  group = NULL,
  mode = "DC",
  which_group_to_be_one = NULL,
  adjust.method = "BH",
  FDRCutoff = 0.05
)
}
\arguments{
\item{mutation_data}{Dataframe or matrix of gene mutation data. Should be a binary data with 0 as the wild type while 1 for mutation.}

\item{group}{Binary data.  Applys to "Separate" or "DC" mode.}

\item{mode}{Which mode to be used. Three modes are included: "DC"(evaluate the significance of differential cooccurrence between two different groups),"Separate"(evaluate the pairwise cooccurrence in each group separately),"Single"(evalute the pairwise cooccurrence for the entire dataset).}

\item{which_group_to_be_one}{One category of the group parameter. Applys to "Separate" or "DC" mode. For "DC" mode, this category is modeled.}

\item{adjust.method}{Method for p value adjustment. Should be one of the "holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr", "none".}

\item{FDRCutoff}{FDR cutoff}
}
\value{
A list with odds ratio, raw p value and fasle discovery rate
}
\description{
Evaluate the cooccurrence or differential cooccurrence of the gene mutation data
}
\examples{
data(gbm_dat)
cohort<-gbm_dat$cohort;unique(cohort)
dat<-gbm_dat[,-1]

#Evaluate pairwise cooccurrence of genomic alteration of the entire gbm_dat data.
Result_Single<-DC.CO_Evaluation(mutation_data=dat,
                                mode="Single",
                                adjust.method="BH",
                                FDRCutoff=0.05)

#Evaluate pairwise cooccurrence for the two cohorts separately.
Result_Separate<-DC.CO_Evaluation(mutation_data=dat,
                                  group=cohort,
                                  which_group_to_be_one="UT",
                                  mode="Separate",
                                  adjust.method="BH",
                                  FDRCutoff=0.05)

#Evaluate the significance of differential cooccurrence between patients with PTEN=0 and PTEN=1
UT_dat<-gbm_dat[gbm_dat$cohort=="UT",-1]
PTEN_grp<-UT_dat$PTEN;table(PTEN_grp)
dat2<-subset(UT_dat,select=-PTEN)
Result_DC<-DC.CO_Evaluation(mutation_data=dat2,
                           group=PTEN_grp,
                           which_group_to_be_one=1,
                           mode="DC",
                           adjust.method="BH",
                           FDRCutoff=0.1)
}
