% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_data.R
\docType{data}
\name{gbm_dat}
\alias{gbm_dat}
\title{gbm_dat}
\usage{
data(gbm_dat)
}
\description{
Genomic alterations data from IDH-WT GBM patients. The dataset consists of 23 genes. A total of 498 samples from 2 different cohort are involved (232 from UThealth at houston and 266 from TCGA cohort).Please refer to the publication for more details.
}
\examples{
data(gbm_dat)
cohort<-gbm_dat$cohort
dat<-gbm_dat[,-1]
}
\references{
Yan et al. (2020) JCO Precision Oncology 4, 575-584
(\href{JCOPO}{https://ascopubs.org/doi/abs/10.1200/PO.19.00385})
}
\keyword{datasets}
