% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stat.Extraction.R
\name{Stat.Extraction}
\alias{Stat.Extraction}
\title{Stat.Extraction}
\usage{
Stat.Extraction(obj = obj_input, WhichGrp_adjP.Separate = NULL)
}
\arguments{
\item{obj}{Object generated by DC.CO_Evaluation function. Required}

\item{WhichGrp_adjP.Separate}{Which group to adjust p value in separate mode. Restricts to separate mode only. Option includes: NULL(p value adjusted by entire dataset), "All"(p value adjusted by entire dataset), "None"(no p value adjustment), "Each"(p value adjusted within each group). The value can also be one of the category of the group parameter. In this case, p value within the specified group is adjusted while the other group shows the raw p value. Optional. Default: NULL}
}
\value{
A data frame with log odds ratio, raw p value, fasle discovery rate and significance
}
\description{
Extract the statistics from the object generated by DC.CO_Evaluation
}
\examples{
data(gbm_dat)
cohort<-gbm_dat$cohort;unique(cohort)
dat<-gbm_dat[,-1]

Result_Separate<-DC.CO_Evaluation(mutation_data=dat,
                                  group=cohort,
                                  which_group_to_be_one="UT",
                                  mode="Separate",
                                  adjust.method="BH",
                                  FDRCutoff=0.05)
Sig_Separate<-Stat.Extraction(obj=Result_Separate)
}
